+++ src/server/settings/settings-router.ts
@@ -1,9 +1,8 @@
 import { router } from "../trpc/trpc"
-import { checkForUpdate, get, save } from "./settings-procedures"
+import { get, save } from "./settings-procedures"
 
 export let settings = router({
     settings: {
-        checkForUpdate,
         get,
         save
     }